Split UI into multiple es6 modules
