+++ js/models/Bookmark.js
@@ -1,0 +1,5 @@
+import Backbone from 'backbone'
+
+export default Backbone.Model.extend({
+  urlRoot: 'bookmark'
+})
