+++ js/models/Bookmarks.js
@@ -1,0 +1,6 @@
+import Backbone from 'backbone'
+
+export default Backbone.Collection.extend({
+  model: Bookmark
+, url: 'bookmark'
+})
