+++ js/models/Tag.js
@@ -1,0 +1,6 @@
+import Backbone from 'backbone'
+
+export default Backbone.Model.extend({
+  idAttribute: 'name'
+, urlRoot: 'tag'
+})
